requirements.txt
# MAI-UI Core Dependencies
Jinja2>=3.1.0
numpy>=2.0.0
openai>=1.0.0
Pillow>=10.0.0

# Web UI Dependencies
gradio>=4.0.0
pyyaml>=6.0.0
